Delegated the work of clone detector output parsers to JDeodorant
CloneToolParser is now a concrete class having the necessary code for
delegation/writing to the excel file. Unnecessary methods have gone from
this class. This also frees us from having dependency on Cedar code.

Note that, in the past, we had repeated clone groups in the excel file,
with the flag "Repeated" in the corresponding column. However, as we
don't allow the repeated clone groups in the parsers in JDeodorant
(i.e., the returned results essentially will not contain the repeated
clone groups), such groups will not be written to the excel files.

This fixes #1.